Flask-Web-App-Tutorial-main/requirements.txt
flask
Flask-SQLAlchemy
flask-login
flask_cors
joblib
requests